requirements.txt
aiohttp==3.9.1
aiofiles==23.2.1
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
websockets==12.0
numpy==1.26.4
httpx>=0.25.2
python-dotenv
azure-identity
azure-ai-resources
Pillow
azure-ai-projects
semantic-kernel
openai
llama-index-core
llama-index-readers-file
llama-index-multi-modal-llms-azure-openai
llama-index-vector-stores-qdrant
llama-index-embeddings-huggingface
qdrant-client
llama-index-llms-azure-openai
llama-index-embeddings-azure-openai
matplotlib
IPython
torch 
torchvision 
torchaudio
docx2txt
mutagen
moviepy
python-pptx